Imported "TPC-C load" program from TPC-C specification version 5.11
This is a literal copy-paste of the code. I am aware that copying this
code from the specification does not preclude us from actually reading
the specification and making sure this code adheres to the
specification. Spcifically, the document says "In case of discrepancy
between the specifications and the programming examples, the
specifications prevail".

I picked up this code, instead of writing one on my own, in C or
JavaScript, just for convenience. I don't see this code as a
performance-critical component of the benchmark execution.