Cannot get the ECPG code to compile; abandoning effort.
Well, that was a perfect waste of 3 hours. There are no good examples of
ECPG usage in the wild. Postgres documentation also agrees that ECPG is
not well documented. Why spend time on something that's not well
exercised in the field.

Abandoning this effort in favor of data loading done in pl/pgsql.